+++ .github/workflows/publish-examples.yml
@@ -12,62 +12,29 @@ jobs:
     env:
       # leave empty for /
       PUBLIC_URL_PREFIX: ""
+      RUSTUP_TOOLCHAIN: nightly
 
     steps:
       - uses: actions/checkout@v2
       - uses: actions-rs/toolchain@v1
         with:
-          toolchain: stable
+          toolchain: nightly
           target: wasm32-unknown-unknown
+          components: rust-src
           override: true
           profile: minimal
+
+      - name: Write optimisation flags
+        run: ./ci/write-optimisation-flags.sh
 
       - uses: Swatinem/rust-cache@v1
 
       - uses: jetli/trunk-action@v0.1.0
         with:
-          # Optional version of trunk to install(eg. 'v0.8.1', 'latest')
           version: 'latest'
 
-
-
       - name: Build examples
-        run: |
-          output="$(pwd)/dist"
-          mkdir "$output"
-
-          for path in examples/*; do
-            if [[ ! -d $path ]]; then
-              continue
-            fi
-
-            example=$(basename "$path")
-
-            # multi_thread doesn't work yet. See <https://github.com/thedodd/trunk/issues/40>.
-            if [[ "$example" == "agents" ]]; then
-              continue
-            fi
-
-            # web_worker_fib does not compile with trunk. See <https://github.com/thedodd/trunk/issues/46>.
-            if [[ "$example" == "web_worker_fib" ]]; then
-              continue
-            fi
-
-            # ssr does not need trunk
-            if [[ "$example" == "simple_ssr" ]]; then
-              continue
-            fi
-
-            echo "building: $example"
-            (
-              cd "$path"
-              dist_dir="$output/$example"
-
-              trunk build --release --dist "$dist_dir" --public-url "$PUBLIC_URL_PREFIX/$example"
-            )
-          done
-
-
+        run: ./ci/build-examples.sh
 
       - name: Deploy to Firebase
         uses: siku2/action-hosting-deploy@v0
