+++ Cargo.toml
@@ -8,7 +8,6 @@ members = [
   "packages/yew-router-macro",
 
   # Examples
-  "examples/agents",
   "examples/boids",
   "examples/contexts",
   "examples/counter",
