+++ examples/README.md
@@ -29,7 +29,6 @@ As an example, check out the TodoMVC example here: <https://examples.yew.rs/todo
 
 | Example                                            | CT | Description                                                                                                                        |
 | -------------------------------------------------- | -- | ---------------------------------------------------------------------------------------------------------------------------------- |
-| [agents](agents)                                   | S  | Cross-component communication using [Agents](https://yew.rs/docs/concepts/agents)                                                  |
 | [boids](boids)                                     | S  | Yew port of [Boids](https://en.wikipedia.org/wiki/Boids)                                                                           |
 | [contexts](contexts)                               | F  | A technical demonstration of Context API.                                                                                          |
 | [counter](counter)                                 | S  | Simple counter which can be incremented and decremented                                                                            |
@@ -40,7 +39,7 @@ As an example, check out the TodoMVC example here: <https://examples.yew.rs/todo
 | [futures](futures)                                 | S  | Demonstrates how you can use futures and async code with Yew. Features a Markdown renderer.                                        |
 | [game_of_life](game_of_life)                       | S  | Implementation of [Conway's Game of Life](https://en.wikipedia.org/wiki/Conway%27s_Game_of_Life)                                   |
 | [inner_html](inner_html)                           | S  | Embeds an external document as raw HTML by manually managing the element                                                           |
-| [js_callback](js_callback)                         | S  | Interacts with JavaScript code                                                                                                     |
+| [js_callback](js_callback)                         | F  | Interacts with JavaScript code                                                                                                     |
 | [keyed_list](keyed_list)                           | S  | Demonstrates how to use keys to improve the performance of lists                                                                   |
 | [mount_point](mount_point)                         | S  | Shows how to mount the root component to a custom element                                                                          |
 | [nested_list](nested_list)                         | S  | Renders a styled list which tracks hover events                                                                                    |
@@ -49,12 +48,12 @@ As an example, check out the TodoMVC example here: <https://examples.yew.rs/todo
 | [portals](portals)                                 | S  | Renders elements into out-of-tree nodes with the help of portals                                                                   |
 | [router](router)                                   | S  | The best yew blog built with `yew-router`                                                                                          |
 | [simple_ssr](simple_ssr)                           | F  | Demonstrates server-side rendering                                                                                                 |
-| [suspense](suspense)                               | F  | This is an example that demonstrates `<Suspense />` support                                                                          |
-| [function_memory_game](function_memory_game)       | F  | Implementation of [Memory Game](https://github.com/bradlygreen/Memory-Game)                                            |
+| [suspense](suspense)                               | F  | This is an example that demonstrates `<Suspense />` support                                                                        |
+| [function_memory_game](function_memory_game)       | F  | Implementation of [Memory Game](https://github.com/bradlygreen/Memory-Game)                                                        |
 | [timer](timer)                                     | S  | Demonstrates the use of the interval and timeout services                                                                          |
 | [todomvc](todomvc)                                 | S  | Implementation of [TodoMVC](http://todomvc.com/)                                                                                   |
 | [two_apps](two_apps)                               | S  | Runs two separate Yew apps which can communicate with each other                                                                   |
-| [web_worker_fib](web_worker_fib)                   | S  | Calculate fibonacci value of a number in a web worker thread                                                                       |
+| [web_worker_fib](web_worker_fib)                   | S  | Calculate fibonacci value of a number in a web worker thread using `gloo-worker`                                                   |
 | [webgl](webgl)                                     | S  | Controls a [WebGL canvas](https://developer.mozilla.org/en-US/docs/Web/API/WebGL_API/Tutorial/Getting_started_with_WebGL) from Yew |
 
 ## Next steps
