--- examples/agents/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "agents"
-version = "0.1.0"
-authors = ["Denis Kolodin <deniskolodin@gmail.com>"]
-edition = "2021"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-log = "0.4"
-serde = { version = "1.0", features = ["derive"] }
-wasm-logger = "0.2"
-yew = { path = "../../packages/yew", features = ["csr"] }
-yew-agent = { path = "../../packages/yew-agent" }
-gloo-timers = "0.2"
