--- examples/agents/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>Yew • Multi-Thread</title>
-
-    <link data-trunk rel="rust" data-bin="app" />
-    <link data-trunk rel="rust-worker" data-bin="worker" />
-  </head>
-
-  <body></body>
-</html>
