--- examples/agents/src/bin/app.rs
@@ -1,4 +1,0 @@
-fn main() {
-    wasm_logger::init(wasm_logger::Config::default());
-    yew::Renderer::<agents::App>::new().render();
-}
