+++ examples/web_worker_fib/index.html
@@ -1,0 +1,13 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+        <title>Yew • Web Worker Fibonacci</title>
+
+        <link data-trunk rel="rust" href="Cargo.toml" data-bin="app" data-type="main" />
+        <link data-trunk rel="rust" href="Cargo.toml" data-bin="worker" data-type="worker" />
+    </head>
+
+    <body>
+    </body>
+</html>
