--- examples/web_worker_fib/serve.sh
@@ -1,9 +1,0 @@
-#!/usr/bin/env bash
-if ! command -v python3 2>&1 >/dev/null;
-then
-	echo 'error: you must install python3'
-	exit
-fi
-
-cd static/
-python3 -m http.server 8080
