+++ examples/web_worker_fib/src/agent.rs
@@ -50,6 +50,6 @@ impl yew_agent::Worker for Worker {
     }
 
     fn name_of_resource() -> &'static str {
-        "wasm.js"
+        "worker.js"
     }
 }
