+++ examples/web_worker_fib/src/bin/app.rs
@@ -1,0 +1,3 @@
+fn main() {
+    yew::Renderer::<yew_worker_fib::App>::new().render();
+}
