+++ examples/web_worker_fib/src/bin/worker.rs
@@ -1,0 +1,6 @@
+use yew_agent::PublicWorker;
+use yew_worker_fib::agent::Worker;
+
+fn main() {
+    Worker::register();
+}
